Move builtin call summaries into pointer_flow
